fix : change enrollment Response name and change Ids types from Integer To Long

+++ src/main/java/com/jonas/TechEventsRegistration/DTO/Enrollment/EnrollmentRequest.java
@@ -14,9 +14,9 @@ import java.time.LocalDate;
 public class EnrollmentRequest {
 
     @NotNull
-    private Integer participantId;
+    private Long participantId;
     @NotNull
-    private Integer eventId;
+    private Long eventId;
 
     @JsonFormat(pattern = "dd/MM" + "/yyyy")
     private LocalDate enrollmentDate;
